fixed hex

+++ libftprintf/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: codespace <codespace@student.42.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/23 12:27:31 by nastamid          #+#    #+#             */
-/*   Updated: 2024/10/01 14:54:29 by codespace        ###   ########.fr       */
+/*   Updated: 2024/10/02 13:14:43 by codespace        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,7 +25,7 @@ int		print_string(char *str);
 int		print_digit(int digit);
 int		print_udigit(unsigned int digit);
 int		print_pointer(void *ptr);
-int		print_hex(long n, bool is_uppercase);
+int		print_hex(unsigned int n, bool is_uppercase);
 int		vft_printf(const char *format, va_list args);
 int		ft_printf(const char *format, ...);
